Revert "Native vs Jgit (#86)"
This reverts commit 463d850b6ff7c93f652331a1ed4ca078a3d61a67.
